feat(news list) finish navigation
