+++ app/src/main/res/values/strings.xml
@@ -2,4 +2,5 @@
     <string name="app_name">News</string>
     <string name="error">Something went wrong, try again!</string>
     <string name="headline_image_description">Image related to headline news</string>
+    <string name="author">"Author: "</string>
 </resources>
