+++ core/network/src/main/java/com/giovannidiluca/network/model/NewsResponse.kt
@@ -8,7 +8,7 @@ data class NewsResponse(
 
 data class ArticleResponse(
     val source: Source,
-    val author: String,
+    val author: String?,
     val title: String,
     val description: String?,
     val url: String,
